+++ tests/specs/exports/2016-06-18.slim.js
@@ -3,13 +3,15 @@
 
 let config = {
   _class: '$config',
-  nextVarStrat: 'naive',
-  next_value_func: 'min',
+  varStratConfig: {
+    type: 'naive',
+  },
+  valueStratName: 'min',
   targetedVars: 'all',
   var_dist_options: {
-    width: {list: [4], distributor_name: 'list'},
-    'width&n=1': {list: [4], distributor_name: 'list'},
-    'width&n=2': {list: [4], distributor_name: 'list'},
+    width: {list: [4], valtype: 'list'},
+    'width&n=1': {list: [4], valtype: 'list'},
+    'width&n=2': {list: [4], valtype: 'list'},
   },
   timeout_callback: undefined,
   all_var_names: [