+++ tests/specs/propagators/markov.spec.js
@@ -24,7 +24,7 @@ describe('propagators/markov.spec', function() {
         id: 'A',
         domain: fixt_arrdom_range(0, 0, true),
         distributeOptions: {
-          distributor_name: 'markov',
+          valtype: 'markov',
           legend: [0],
           matrix: [
             {vector: [1]},
@@ -45,7 +45,7 @@ describe('propagators/markov.spec', function() {
         id: 'A',
         domain: fixt_arrdom_range(0, 0, true),
         distributeOptions: {
-          distributor_name: 'markov',
+          valtype: 'markov',
           legend: [1],
           matrix: [
             {vector: [1]},
@@ -68,7 +68,7 @@ describe('propagators/markov.spec', function() {
           id: 'A',
           domain: fixt_arrdom_range(0, 0, true),
           distributeOptions: {
-            distributor_name: 'markov',
+            valtype: 'markov',
             legend: [0],
             matrix: [
               {vector: [0]},
@@ -89,7 +89,7 @@ describe('propagators/markov.spec', function() {
           id: 'A',
           domain: fixt_arrdom_range(0, 0, true),
           distributeOptions: {
-            distributor_name: 'markov',
+            valtype: 'markov',
             legend: [0],
             matrix: [
               {vector: [0]},
@@ -113,7 +113,7 @@ describe('propagators/markov.spec', function() {
           id: 'A',
           domain: fixt_arrdom_range(0, 0, true),
           distributeOptions: {
-            distributor_name: 'markov',
+            valtype: 'markov',
             legend: [0],
             matrix: [{
               vector: [0],
@@ -139,7 +139,7 @@ describe('propagators/markov.spec', function() {
           id: 'A',
           domain: fixt_arrdom_range(0, 0, true),
           distributeOptions: {
-            distributor_name: 'markov',
+            valtype: 'markov',
             legend: [0],
             matrix: [{
               vector: [1],