+++ tests/specs/solver.markov.spec.js
@@ -382,7 +382,7 @@ describe('solver.markov.spec', function() {
         id: 'B_MARK',
         domain: fixt_arrdom_range(0, 1, true),
         distributeOptions: {
-          distributor_name: 'markov',
+          valtype: 'markov',
           legend: [2],
           matrix: [
             {vector: [1]},
@@ -398,8 +398,8 @@ describe('solver.markov.spec', function() {
         max: 1,
         distribute: {
           // distribute should be ignored as it should reject immediately
-          var: 'throw',
-          val: 'throw',
+          varStrategy: {type: 'throw'},
+          valueStrategy: 'throw',
         },
       });